+++ src/extension.ts
@@ -1,11 +1,14 @@
 import * as vscode from "vscode";
-import * as playdate from "./playdate";
+import { compileWithDefaults, runSimulatorWithDefaults } from "./command";
 
 export function activate(context: vscode.ExtensionContext) {
-	const disposable = vscode.commands.registerCommand('extension.runSimulator', () => {
-        playdate.compile() && playdate.runSimulator();
-	});
+  const disposable = vscode.commands.registerCommand(
+    "extension.runSimulator",
+    () => {
+      compileWithDefaults() && runSimulatorWithDefaults();
+    }
+  );
 
-	context.subscriptions.push(disposable);
+  context.subscriptions.push(disposable);
 }
 export function deactivate() {}