+++ src/startSim.ts
@@ -1,0 +1,4 @@
+import { MockDebugSession } from "./MockDebugSession";
+
+// @ts-ignore
+MockDebugSession.run(MockDebugSession);