+++ CHANGELOG.md
@@ -2,12 +2,11 @@
 
 ## [Unreleased]
 
-- convert plugin from debugger to command
+- add support to run from command palette
 
 ## [0.0.5]
 
 - add support for Winodws
-
 
 ## [0.0.1]